change short to byte in Colouring

+++ printing-house-server/repo/src/main/kotlin/pl/macia/printinghouse/server/bmodel/Colouring.kt
@@ -1,7 +1,7 @@
 package pl.macia.printinghouse.server.bmodel
 
 sealed interface Colouring {
-    var colouringId: Short?
+    var colouringId: Byte?
     var secondSide: Byte
     var firstSide: Byte
 }
